backend/requirements.txt
flask
sentence-transformers
faiss-cpu
